requirements.txt
python==3.10.7
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0